refactor: move constants from types to constants file
