+++ src/components/AudioSpeedControl.tsx
@@ -1,7 +1,8 @@
 "use client";
 import { useEffect } from "react";
 
-import { EditRecipe, SPEED_STEPS } from "@/lib/types";
+import { EditRecipe } from '@/lib/types'
+import { SPEED_STEPS } from '@/lib/constants'
 import { Volume2, VolumeX, Gauge } from "lucide-react";
 import { cn } from "@/lib/utils";
